+++ state-ex-machina-ext-compose/src/commonMain/kotlin/state/ex/machina/compose/ModelExtensions.kt
@@ -24,16 +24,19 @@ private fun <I : Intent> intentsFlowForCompose(
     return intentCallback to intentFlow
 }
 
-data class MviComponents<S : State, I : Intent>(
-    val state: S,
+interface StateMachineComposeContract<S : State, I : Intent> {
+    val state: S
     val onIntent: (I) -> Unit
-)
+
+    operator fun component1(): S = state
+    operator fun component2(): (I) -> Unit = onIntent
+}
 
 @Composable
-fun <S : State, I : Intent> rememberMviComponent(
+fun <S : State, I : Intent> rememberStateMachine(
     model: Model<S, I>,
     scope: CoroutineScope = rememberCoroutineScope()
-): MviComponents<S, I> {
+): StateMachineComposeContract<S, I> {
     val (stateFlow, onIntent) = remember {
         val (onIntent, intentFlow) = intentsFlowForCompose<I>(scope)
         val stateFlow = model.subscribeTo(intentFlow)
@@ -41,8 +44,8 @@ fun <S : State, I : Intent> rememberMviComponent(
     }
     val state = stateFlow.collectAsState()
 
-    return MviComponents(
-        state = state.value,
-        onIntent = onIntent
-    )
+    return object : StateMachineComposeContract<S, I> {
+        override val state: S = state.value
+        override val onIntent = onIntent
+    }
 }