+++ state-ex-machina-foundation/src/commonMain/kotlin/state/ex/machina/dsl/StateMachine.kt
@@ -12,14 +12,6 @@ import state.ex.machina.foundation.Intent
 import state.ex.machina.foundation.Reducer
 import state.ex.machina.foundation.State
 import state.ex.machina.foundation.Store
-import state.ex.machina.foundation.store
-
-fun <S : State, I : Intent> stateMachine(
-    initialState: S,
-    intents: Flow<I>,
-    coroutineScope: CoroutineScope,
-    builder: IntentStateMachineScope<S, I>.() -> Unit
-) = stateMachine(store(initialState), intents, coroutineScope, builder)
 
 fun <S : State, I : Intent> stateMachine(
     store: Store<S>,
@@ -36,38 +28,47 @@ fun <S : State, I : Intent> stateMachine(
     return store.stateFlow
 }
 
-class StateMachineBuilder<S : State, I : Intent>(
-    private val store: Store<S>,
+private class StateMachineBuilder<S : State, I : Intent>(
+    store: Store<S>,
     intents: Flow<I>
-) : IntentStateMachineScope<S, I>(intents), SideEffectScope<S> {
-
+) : IntentStateMachineScope<S, I>(intents) {
+    private val sideEffectScope = StoreSideEffectScope(store)
     private val _actionHandlers = mutableListOf<Flow<*>>()
+
     val actionHandlers: List<Flow<*>>
         get() = _actionHandlers.toList()
 
+    override fun <D> on(flow: Flow<D>) = ActionHandler(
+        sideEffectScope = sideEffectScope,
+        actionFlow = flow,
+        onActionHandledFlow = { _actionHandlers += it }
+    )
+}
+
+private class StoreSideEffectScope<S : State>(
+    private val store: Store<S>
+) : SideEffectScope<S> {
     override val currentState: S by store::currentState
     override fun updateState(reducer: Reducer<S>): S = store.update(reducer)
+}
 
-    override fun <D> on(flow: Flow<D>) = ActionHandler(flow)
+private class ActionHandler<S : State, D>(
+    private val sideEffectScope: SideEffectScope<S>,
+    private val actionFlow: Flow<D>,
+    private val onActionHandledFlow: (Flow<*>) -> Unit
+) : ActionScope<S, D> {
+    override fun updateState(
+        reducerBuilder: (D) -> Reducer<S>
+    ) = sideEffect { updateState(reducerBuilder(it)) }
 
-    inner class ActionHandler<D>(
-        val flow: Flow<D>
-    ) : ActionScope<S, D> {
-        override fun updateState(
-            reducerBuilder: (D) -> Reducer<S>
-        ) = sideEffect { updateState(reducerBuilder(it)) }
-
-        @OptIn(ExperimentalCoroutinesApi::class)
-        override fun sideEffect(
-            block: suspend SideEffectScope<S>.(D) -> Unit
-        ) = rawSideEffect { flow ->
-            flow.flatMapMerge { flow<Nothing> { block(it) } }
-        }
-
-        override fun rawSideEffect(
-            block: SideEffectScope<S>.(Flow<D>) -> Flow<*>
-        ) {
-            _actionHandlers += block(flow)
-        }
+    @OptIn(ExperimentalCoroutinesApi::class)
+    override fun sideEffect(
+        block: suspend SideEffectScope<S>.(D) -> Unit
+    ) = rawSideEffect { flow ->
+        flow.flatMapMerge { flow<Nothing> { block(it) } }
     }
+
+    override fun rawSideEffect(
+        block: SideEffectScope<S>.(Flow<D>) -> Flow<*>
+    ) = onActionHandledFlow(sideEffectScope.block(actionFlow))
 }